ssm: Delete ssm parameter based on resource name
Use the `resource_name` to delete the SSM Parameter when using Cloudformation.

Will fallback to using the Cloudformation JSON if `resource_name` is not
specified.

The JSON `cloudformation_json` would not match the resource name if
Parameter values are being used in the Cloudformation template to
represent the SSM ParameterStore name.

In the `cloudformation_json` case below, the SSM parameter that is
passed to `ssm_backend.delete_parameter` is the literal value, not the
parsed value.

```
{
  "Properties": {
    "Name": {
      "Fn::Sub": "${Name}"
    },
    "Type": "String",
    "Value": {
      "Ref": "Value"
    }
  },
  "Type": "AWS::SSM::Parameter"
}

```